Update to Version 2.0

+++ Programs/virtualos-manager.lua
@@ -1,20 +1,41 @@
 local sDir = "/var/virtualos-manager"
-local sSaveFile = "list.lua"
+local sSaveFile = "drivelist.lua"
 local sVirtualPath
 local bSelect = false
 local tList = {}
 local nSelect
 local w,h = term.getSize()
+local sVersion = "2.0"
 
 if fs.exists(fs.combine(fs.getDir(shell.getRunningProgram()),"virtualos.lua")) then
-    sVirtualPath = fs.combine(fs.getDir(shell.getRunningProgram()),"virtualos.lua")
+    sVirtualPath = "/"..fs.combine(fs.getDir(shell.getRunningProgram()),"virtualos.lua")
 elseif fs.exists("/usr/bin/virtualos.lua") then
     sVirtualPath = "/usr/bin/virtualos.lua"
 else
     error("Can't find VirtualOS",0)
 end
 
+local function showText(sText)
+w,h = term.getSize()
+term.setBackgroundColor(colors.white)
+term.clear()
+term.setTextColor(colors.black)
+term.setCursorPos(1,1)
+print(sText)
+term.setBackgroundColor(colors.blue)
+term.setCursorPos(1,h)
+term.clearLine()
+term.write("OK")
+while true do
+    local ev,me,x,y = os.pullEvent("mouse_click")
+    if y == h then
+        return
+    end
+end
+end
+
 local function redrawMenu()
+w,h = term.getSize()
 term.setBackgroundColor(colors.white)
 term.clear()
 term.setBackgroundColor(colors.blue)
@@ -22,9 +43,9 @@ term.setCursorPos(1,1)
 term.clearLine()
 term.setTextColor(colors.black)
 if bSelect then
-    term.write("Run Delete New Link")
+    term.write("Run Delete New Link About")
 else
-    term.write("New")
+    term.write("New About")
 end
 term.setCursorPos(w,1)
 term.blit("X","f","e")
@@ -51,9 +72,12 @@ if os.version():sub(9) == tList[nSelect]["version"] then
 else
     sRomPath = fs.combine(sDir,"rom/"..tList[nSelect]["version"])
 end
-local sExtra = ""
-if tList[nSelect]["version"] == "1.5" then
-    sExtra = " --oldnative"
+local sExtra = " \"--title=CraftOS "..tList[nSelect]["version"].."\""
+if tonumber(tList[nSelect]["version"]) <= 1.5 then
+    sExtra = sExtra.." --oldnative"
+end
+if tList[nSelect]["version"] == "1.2" then
+    sExtra = sExtra.." --diskapi"
 end
 return sVirtualPath.." --biosPath="..fs.combine(sDir,"bios/"..tList[nSelect]["version"])..".lua --romPath="..sRomPath.." --rootPath="..fs.combine(sDir,"drive/"..tList[nSelect]["name"])..sExtra
 end
@@ -71,7 +95,7 @@ return true
 end
 
 local function getVersion()
-local tVersion = {"1.5","1.6","1.7","1.8"}
+local tVersion = {"1.0","1.1","1.2","1.3","1.4","1.5","1.6","1.7","1.8"}
 term.setBackgroundColor(colors.white)
 term.clear()
 term.setBackgroundColor(colors.blue)
@@ -94,6 +118,11 @@ end
 
 local function getBios(sVersion)
 local tBios = {}
+tBios["1.0"] = "https://raw.githubusercontent.com/Wilma456/ComputercraftRom/1.0/bios.lua"
+tBios["1.1"] = "https://raw.githubusercontent.com/Wilma456/ComputercraftRom/1.1/bios.lua"
+tBios["1.2"] = "https://raw.githubusercontent.com/Wilma456/ComputercraftRom/1.2/bios.lua"
+tBios["1.3"] = "https://raw.githubusercontent.com/Wilma456/ComputercraftRom/1.3/bios.lua"
+tBios["1.4"] = "https://raw.githubusercontent.com/Wilma456/ComputercraftRom/1.4/bios.lua"
 tBios["1.5"] = "https://raw.githubusercontent.com/alekso56/ComputercraftLua/1.5/bios.lua"
 tBios["1.6"] = "https://raw.githubusercontent.com/alekso56/ComputercraftLua/1.65/bios.lua"
 tBios["1.7"] = "https://raw.githubusercontent.com/alekso56/ComputercraftLua/1.79/bios.lua"
@@ -133,6 +162,22 @@ term.write("Please choose a name:")
 term.setBackgroundColor(colors.white)
 term.setCursorPos(1,2)
 local name = read()
+if name == "" then
+    showText("The Name can't be empty")
+    redrawMenu()
+    return
+elseif name:find(" ") ~= nil then
+    showText("No spaces allowed")
+    redrawMenu()
+    return
+end
+for k,v in ipairs(tList) do
+    if v.name == name then
+        showText("This Name does already exists")
+        redrawMenu()
+        return
+    end
+end
 local ver = getVersion()
 getBios(ver)
 getRom(ver)
@@ -156,9 +201,32 @@ term.write("Please enter Path:")
 term.setBackgroundColor(colors.white)
 term.setCursorPos(1,2)
 local path = read()
+if fs.exists(path) then
+    term.clear()
+    term.setCursorPos(1,1)
+    term.write("File exists. Overwrite? (Y/N)")
+    while true do
+        local ev,me = os.pullEvent("key")
+        if me == keys.y then
+            break
+        elseif me == keys.n then
+            redrawMenu()
+            return
+        end
+    end
+end
 local file = fs.open(path,"w")
-file.write('shell.run("'..getCommand()..'")')
+file.write('shell.run("'..getCommand():gsub('"','\\"')..'")')
 file.close()
+redrawMenu()
+end
+
+local function about()
+showText("VirtualOS-Manager Version "..sVersion..[[ made by Wilma456
+
+This is a GUI for VirtualOS, which allows you to run CraftOS in a virtual Machine.
+
+VirtualOS and VirtualOS-Manager are both licensed under the BSD 2-clause "Simplified" License]])
 redrawMenu()
 end
 
@@ -179,6 +247,8 @@ while true do
                 if x < 4 then
                     newMachine()
                     redrawMenu()
+                elseif x > 4 and x < 10 then
+                    about()
                 elseif x == w then
                     break
                 end
@@ -187,8 +257,6 @@ while true do
                     term.clear()
                     term.setCursorPos(1,1)
                     if shell.run(getCommand()) == false then
-                        term.setTextColor(colors.black)
-                        term.setBackgroundColor(colors.white)
                         print("Press any Key to Continue")
                         os.pullEvent("key")
                     end
@@ -206,6 +274,8 @@ while true do
                     redrawMenu()
                 elseif x > 15 and x < 20 then
                     newLink()
+                elseif x > 20 and x < 26 then
+                    about()
                 elseif x == w then
                     break
                 end
